Make the document exporter render option labels
We need the option labels to be included in the artefact presented by
the content api to the specialist-frontend so that the correct labels
appear in the page.

Rather than making the content api communicate with the finder api,
introducing an additional dependency for the finder api, we keep that
knowledge encapsulated in the specialist publisher by rendering the
option labels as well as the option values into the rendered document.