Allow SpecialistDocument to be pinned to a particular version
This adds a #published_version method to specialist_document which
creates an instance of the SpecialistDocument pinned to the version
number of the published version.

It also adds a keyword argument to the constructor to accept the version
number that the SpecialistDocument is pinned to.

We introduce the concept of 'exposed_edition' which points to the
edition with the pinned version.

Also added a publish! method to encapsulate some of the state change
functionality which was being performed in the
SpecialistDocumentRepository.

The specs are changed to be more explicit about version numbers as the
edition version number now is important in the behaviour of the
specialist document.