Avoid title/slug clashes in created cases
If this method was invoked twice (eg. to create draft and published
cases) then the numbering of the titles would start from 1 on each
invocation and therefore the titles would overlap. This would in turn
mean that slugs would overlap. We don't have any enforcement of slugs in
the models, only in panopticon, so this these editions would've been
created successfully with duplicate slugs. Clearly not desirable.

This change ensures that the title is unique across all invocations.