Use #save! when saving document editions
If the document fails to save for any reason we want the flow to stop so
that we don't end up registering non-existent documents with Panopticon.

The document is validated in advance of calling #save!, so if it fails
it should be for an exceptional reason.

Making this change required setting document_id on new editions before
they are passed to the repository so that they pass validation.