Remove render_with method
It provided a little bit of syntactic sugar at the cost of making
controllers and views harder to reason about for newcomers to the
codebase, especially for frontend developers who may not be aware of
where to look for the render_with definition. These developers will be
more familiar with the `render(…, locals: {…})` syntax because it's
commonly used in views to render partials.