Fix for preview on new documents
There was an issue again where #preview wouldn't work with new
documents because it was passing an id which didn't exist, so it was
404ing and redirected by the rescue in the application_controller. In
order to fix this, we've added another route which the preview button
uses for new documents instead. Also added a test to stop this
regressing again.