Allowing users to attach files to AAIB Reports
This commit adds an AAIB Reports Attachments controller and uses the
same views, with the only change being paths for partials and a switch
for the breadcrumbs in the attachment pages.

This also refactors out some steps to be a bit more generic and adds a
cucumber test for the AAIb Reports Attachments workflow.