Give Document formatters specific "kind"
Instead of being generic about how we register documents with panopticon give them an explicit "kind" which is stored with the PanopticonMapping as resource_type. In order to find these again we will need to change the PanopticonMapping#documents scope to pull all types of specialist-document.

While we are here lets also rename DocumentArtefactFormatter to a CmaCaseArtefactFormatter which is its correct name!