Rename misleading observers
Some of the methods in the `ObserverRegistry` and the wiring suggest
they provide observers for generic documents but actually provide
observers for CMA cases specifically.

- document_publication => cma_case_publication
- document_update => cma_case_update
- specialist_document_creation_observers => cma_case_creation_observers
- specialist_document_withdrawal_observers => cma_case_withdrawal_observers

Also update two of the `ServiceRegistry` dependencies which revealed the
issue that the `withdraw_aaib_report` service was using observers for
withdrawal of CMA cases. That's been fixed too.