Rename misleading repositories
Rename `document_repository` to `cma_case_repository` in the
`ServiceRegistry` and rename `specialist_document_repository` to
`cma_case_repository` in the wiring and throughout several scripts and
tests.

Doing so has revealed a few issues where we might be using the wrong
repository (in `validated_manual_document_factory_factory`) or an overly
specific repository (in the scripts and tests).