Don't pollute SpecialistDocument with errors and changes
No need to pollute the SpecialistDocument here with ActiveModel methods
these shouldn't be required here and should be the responsibility of the
validators.

This does however break the importer as we don't provide 'valid?' or
'errors' methods for the importer because we do not use the
SpecialistDocumentValidator. In order to solve this we use a
NullValidator to implement sensible defaults for these.