Use rails-esque SafeHtmlValidator
Instead of using the govuk_content_models validator begin creating a validator which only validates specific fields in a more railsy way. This validator interacts directly with Govspeak::HtmlValidator driectly and only validates the html on specific fields. It means that our validator now only depends on the Govspeak gem and allows us to remove dependencies on ActiveModel::Dirty which the current gem validator expects. This is much more flexible.

We also remove the need for an SpecialistDocumentValidator to hide away the horrible hacks around 'valid?' and means that we no longer need to try DRY up the document validators and can instead explicitly validate fields for each document type.

This validator can eventually be pulled into the govuk_content_models gem and replace the SafeHtml validator there. We will not do this as part of this set of work in order to contain the large number of changes that would be required by other Gov.uk apps in order for this to work.