Move html rendering to erb template
Fix a bug where we were calling content_tag in the service layer. Instead allow the service to pass the document down to the controller. The controller can now decide which html to render (either the document renderer body or the error html) rather than the service making this decision.

This also means the service no longer needs access to content_tag and ActionView methods.