Remove PublishLogger in favour of PublishingTasks
Instead of creating new log entries for every state of publishing, instead we are going to use publishing tasks as a way of telling the workers which manuals to publish.

There is no need to interact with the workers through a registry as we gain very little from this interface. Instead we can simply pass the worker class directly to the service and interact with it directly.

This means that PublishingLogs can now relate to a successful PublishingTask which will contain state of documents current publication.