Correctly generate change notes from publication log
When a manual and its documents are published, we create a
PublicationLog entry -- one for each document, each time the manual is
published. This is then used to generate the change notes.

However, we only want to generate a change note for a document for the
first PublicationLog entry for each document version.

Filter out only the appropriate entries.