Remove all PublicationLog's from production
At the time of writing (and for the forseeable future...) there are no
published manuals on production.

However, there are a couple of manuals that have been accidentally
published and manually unpublished, which has generated PublicationLog
entries.

Given that we've now changed the format of PublicationLog, adding a
version_number, and we'd like all PublicationLogs to include this piece
of information, destroy all existing ones so that when manuals are
published for real, all PublicationLogs will conform to the expected
schema.