Make sure version number is present
When we interact with PublicationLog we should make sure we alwayds have a version number amnd, for that matter, a slug as this will cause issues in the future.

This was being caught by Sidekiq when queueing publishing. It will also require migrations to be run in order to remove exisiting PublicationLogs again as there is some existing data we are aware of which currently breaks preview.