DRY up common form partials for js preview
We also DRY up the path helpers for new and persisted previews. This involves a slight change to the path for new previews. As rails by default checks persisted to decide wether to build the path on the collection or as a member, I have pluralised the new paths. This matches closer to rails' standards.