Tidy up publish task API by collapsing scopes
As the scope is "for_manual" and the use case of these task logs is such that
they must be in newest first order, this scope now implicitly orders the
returned records.

The separate ordering scope has been removed.

This change makes the API smaller and more testable as chaining is no longer
necessary.