Update SafeHtmlValidator to match content_models
Updated the validator to match the SafeHtml validator used in govuk_content_models this should be pushed to govuk_content_models in order to remove the duplication although it will create a breaking change. The main difference here is that the version in govuk_content_models validates all fields on an object no simply a specified field. This works much more like a railsesque validator and only validates a single field.