Reraise panopticon publication errors for worker
Re raise any errors created by the PanopticonRegisterer in order to catch them in the worker. This is something we would like to change. wrapper classes around the gds-api-adapters should handles errors and error logging themselves in the future. There is thought towards allowing the API wrappers to an object which stores the state of the publishing (PublishingTask maybe?) which could be handed to the worker in order to query the state of a particular publishing job. This would allow us to correctly raise an error in the worker without frivolously catching GdsApi exceptions and making the worker unaware off GdsApi.