Stop exporting manual sections to Panopticon
Only the manual itself needs to be exported to Panopticon for the
purposes of tagging with Organisation and Specialist Topics which allows
it to be included in specialist browse pages.

We have to continue exporting manual change notes for now in order to
register their route. When we changed manuals to be exported to the
Content Store instead of Content API we moved change notes to be part of
the details hash of the manual itself, rather than being a stand alone
document. That means Content Store isn't registering the route for the
change notes. Leaving the Panopticon registration in is a workaround and
should be addressed in the future.