Remove section from the style guide manual
A section has inadvertently been added to the style guide manual which
needs removing. We don't have the facility to unpublish manuals yet, nor
remove them from the (recently added) content store, so we're going to:

1. Remove the manual section from the manual in specialist-publisher
2. Republish the manual manually (heh).

This will make the manual section disappear from the table of contents and
specialist publisher, but it will still show up in site search.