Add 'reform' gem to represent search form
In order to use the rails form helpers, and get niceties such as the
form reflecting the model during the search process, we need the object
storing the form parameters to conform to an ActiveModel style API.

By adding `reform`, we have the ability to do this without backing it
with a persisted model.

In specialist-publisher, we used to do this with app/view_adapters, but
these have become so customised to specialist documents that creating a
new view adapter would be overcomplex for this requirement.