Remove `IdGenerator` class and object-constant-based calls of it
The `IdGenerator` object did very little. It can be summed up as the
following bit of code: `SecureRandom.uuid`. Keeping it around doesn't
add anything and it makes it hard to understand why the constant form
of the object is being called in the way it does.

Removing it means there's one less thing to understand in terms of
using the already existing `*Builder` classes.