Don't validate using `SafeHtml` from govuk_content_models
Keeping this in triggers an error where an uninitialised constant --
`GOVSPEAK_FIELDS` -- can't be found. After some discussion with @alext
we decided to remove this functionality as we probably don't want to
allow embedding HTML in `Attachment` objects.

There's also a (cucumber) test to prevent injection attacks so we know
doing this is safe.

The error as it appears when this is kept in:

`uninitialized constant Attachment::GOVSPEAK_FIELDS (NameError)`