Wrap Finder formats in a Dropdown if GDS Editor
As GDS Editors, the amount of Finders was breaking the header
navigation. Now, we wrap the Finders in dropdown tags if the user is
from GDS. We also display the name of the current finder in the dropdown
if one is selected.

Also, adding a new route and permission check in the view requires
adding a new item in the hash in the application controller.