Make edit_[:document_type] method more generic
Editing a document is requires nothing different for each document format. It simply fills in fields, saves and publishes them. By moving the step to go to a particular page we can make this method much simpler. This will allow us to have more generic methods accross the board removing a massive amount of duplication and making the cucumber steps much easier to work with.