Allow CmaCases to be edited with change notes
As with Manuals CmaCases can now be edited with change notes.
Currently these change notes will only be stored against the
SpecialistDocumentEdition and do not create a PublicationLog entry.
 Which currently means change notes will not be able to be used on
the frontend yet. This would require some work in order to make the
PublicationLog work with other formats other than manuals.