Hide attachments list on update action of document
Bug: go to edit a specialist document which has attachments, submit an
invalid edit, receive 500 error.

This is caused by a similar problem to 4ce549e, but I have resolved it
differently.

Because we aren’t really editing an existing document edition, but
duplicating it and saving that one, the embedded attachments respond to
`#new_record?` with true. This means that url_for(attachment) links to
the plural not the singular route of attachment.

However, for this problem, it is much less important that we display
this correctly on the update action, so instead we simply remove the
ability to edit attachments on an invalid document.

Same concerns as the other commit.