Send Change History to Content API
With the change notes, we were only persisting the latest change note.
This commit inserts a hash of the timestamp and change note into a
change history Array and sends this to the content API.

It also deals with minor update checkbox and major update change note
persistence between versions.