Minore updates dont change last_update in rummager
When exporting to rummager we should not update the last_update
timestamp if the change is minor. This will mean that minor updates do
not change the display order of documents.

Manuals do not respond to #minor_update? and therefore we must
use a subclass for specialist documents explicitly.

I have also added tests for all formats using the shared rspec examples.