Better slug unique warning for Documents
As editors may be editing documents with the same title we needed to
improve the functionality of this warning. This commit changes the
warning text to an alert to be more visible, disables the Publish button
if it is not unique and only makes the uniqueness check happen if the
document isn't currently Published. This means that you can still
Publish already published documents.