Use the correct published_at date
When publishing, we first update the document to say that it is
'published', then we export it to the content api. Because by the
time the exporter calls the published_edition after the document has
been updated to published, this returns the latest_edition instead.

Instead we must find the last major edition and use its updated_at
to retain the published at date across minor updates.