Store public_updated_at for sorting finders and frontend updated_at
As the repository only pulls out the last 2 document editions there is no
garauntee that we will always have pulled out a document edition that is a
major update. Instead we will create a field which stores the date of publish
for the last major update. This is persisted across editions and exported to
rummager and content-api as the date which defines when a document was
last updated from the front end user's perpective.