Make sure to touch public_updated_at on bulk publish
This seems wrong but means we can order things before publishing
in order for rummager to index the items in the correct order.

We should find a better solution for this after the imports as updating
public_updated_at when bulk publishing happens is wrong. This means the
user facing public_updated_at is changed each time a bulk publish is run even
though we do not expect bulk publishes to be a user facing change.