Add metadata and schemas
As we want to remove Finder API from the stack we need to replace the
Specialist Publisher schemas with those from Finder API along with the
metadata files. Because of some of the difference in naming conventions
we need to name the schemas the same as what they're known internally in
Specialist Publisher.

This commit moves them into a finders/#{thing} folder and renames the
ones required.