Always treat multi-select values as Arrays
As we now use the schemaes that are used on the Frontend we need to
treat what previously was single-selects on the Frontend as a
multi-select.

This doesn't have a lot of impact apart from the value to label
conversion when showing a document. This commit always casts the value
to Array as even though the value from the document is a string it is
being converted as a multi select.