Handle bad arguments in #reorder_documents
Be a bit more explicit in the handling of bad arguments. Without this
check it would provide a misleading ArgumentError if `document_order`
was missing an ID, and would silently accept extra or duplicate IDs in
the input.

Also adds a unit test for the base case to check it does the reordering.