Fix sections in How to Publish on GOV.UK manual
Until now it's not been possible for editors to reorder sections in a
manual. It's still not possible for editors to remove sections from a
manual.

To work around these limitations, editors have been copy-pasting content
around, reusing old sections to get the correct ordering. Unfortunately,
the slug of a section is never updated after it's been published, so
many of the slugs are now very wrong. The version history for these
sections is also wrong because it includes content from different
sections.

This migration updates the data but does not republish the manual. That
needs to be done by an editor as a manual task.

See also: https://github.com/alphagov/specialist-publisher/pull/402