Script to publish some outstanding MAIB reports
We imported historic MAIB reports and published them in bulk on January
23rd 2015. Since then, MAIB have found 12 more reports which should have
been imported and have added those reports into the publishing tool.
This script publishes those 12 reports.

Specialist Publisher has no support for specifying a "first published
date" for content that pre-dates GOV.UK. The [finder for MAIB
reports][1], shows the reports in order of time of last update, with the
most recent at the top. In order to have the reports appear in the
correct order, they were published in chronological order.

In order to publish these new reports into the correct places in that
list, this script uses the `date_of_occurrence` and `public_updated_at`
values for the already published reports to find a suitable timestamp to
use.

Once this script has been run on production it can be removed.

[1]: https://www.gov.uk/maib-reports