Implement #content_api_withdrawer once only
Each specialist document observers registry was defining this method in
the exact same way. Move the definition into the abstract superclass and
pull it out of the wiring and subclasses entirely.

If any future subclass needs to vary this behaviour it will still be
easy for that to be done.