Refactor Rummager exporter/withdrawer observers
The only way each method implementation varied was to customise the
formatter class used. Defining a method on each subclass to format a
document for indexing allows us to define `rummager_exporter` and
`rummager_indexer` once on the superclass.

This still leaves each subclass implementation of
`format_document_for_indexing` having to know to wrap the document in
`MarkdownAttachmentProcessor` before formatting it, but this is a step
forward despite that.