Migration to clean up immigration rules manuals
There are three copies of the immigration rules manual in Specialist
Publisher. Two of those are test copies which need to be deleted. One of
those test manuals was unintentionally published a few months ago. The
data about that manual was removed from most of the services that knew
about it, but it was kept in Specialist Publisher.

This migration deletes the two test manuals and updates the slugs of the
third manual and all its sections to use the slug
`guidance/immigration-rules`, originally used by the unintentionally
published version.

This also cleans up any `PublicationLog` records if necessary, but I
don't think there are any to be cleaned up.

There are a few artefacts left around from old manuals (both draft and
published) in Panopticon which could be cleaned up, but I don't think
they will cause any problems with these changes.