First pass at importing legacy CMA cases
This loads the scraped content from JSON files and passes it directly
into a `CreateDocumentService` for CMA cases.

We still need to massage the source data to match the format the
document builder is expecting (ie get the metadata into the right place)
and we still need to get attachments attached.