Import attachments as part of CMA case import
Loads the assets referenced in the JSON files from the scrape, and adds
them as attachments to the imported CMA case. Any links in the body to
the original file are replaced with a Markdown shortcode for the newly
added attachment.

If a document has any attachments which aren't linked to from the body
content, we still add them as attachments but mark that document as a
failed entry in the import log with the filenames of the unlinked
attachments for reference. Note: the "failed" documents still get
imported, so they can be manually tidied up by a content editor using
Specialist Publisher.

Several hundred of the JSON files from the scrape do not specify a
document body. Because this code expects the body to exist, these
documents show up as an error in the import log. They are actually still
persisted with a nil body, which is something we should fix.

Two of the JSON files from the scrape refer to assets that aren't
included in the scrape data -- these show up in the import log as an
error. The document is still persisted, but some of the attachments may
be missing.