Improve logging for CMA import
Previously we marked a row as a failure and put messages about the
record into the logs via the `errors` method. This was misleading
because in all of these cases the document was actually successfully
imported, and the errors are really just notes or caveats about the
import.

Instead, log the document itself by calling `to_s` and wrap the document
in a decorator whenever we want to add extra information to the log
output.