Service and Exporter for withdrawing a manual
No work has been done on the UI for this yet, but this will allow a script to
withdraw a manual.

If the service successfully changes the Manual's state to "withdrawn", it will
persist the change in Specialist Publisher and notify PublishingApi, Rummager
and Panopticon.

Only a published manual can be withdrawn