Replace HTML attachment links in imported body
We fix incomplete bodies by appending the contents of the final report
and appendices and glossary section, those sections contain HTML markup
for tables including links to the attachments.

In order to replace the links to the original attachments with links to
the imported attachments we need to update the regex to match HTML
anchor elements as well as markdown links.

This new regex has two captures, only one of which will apply: either it
will be a markdown link, or an HTML link. This means the captures array
will be either `[<match>, nil]` or `[nil, <match>]`, so `compact` it to
get rid of the `nil` and take the remaining element.