Add CmaImportBodyFixer to fix incomplete bodies
Some of the data that was scraped as part of the work to import legacy
CMA cases has `markup_sections` which aren't included in the `body`
field.

In some cases the missing content are the links to many of the
supporting attachments. That is the case where we are missing the
`final-report-and-appendices-glossary` section and no other sections.

This code checks if that's the case, and appends it to the body. It also
adds some import notes for the log output to list the missing sections
and whether it was fixed automatically.