Handle select2 fields in fill_in_field helper
When a select2 field is initialised it copies the label tag for the
original field, so there are two labels with the same text. That causes
the `find_field` helper to raise a `Capybara::Ambiguous` error.

To work around this, first try to find a select field and if found
select the value from that. If there isn't a select field, just use the
`fill_in` helper and let Capybara work out which field to use.

When select2 is involved, this will find the original select field
bypassing the select2 functionality. Ideally this method would account
for select2 fields and interact with them in the same way a user would
(by clicking to open a dropdown list, then clicking an item). That seems
like a piece of yak-shaving too far for now, so this at least keeps
everything working.