Allow multiple market_sector values for CMA cases
Since `market_sector` is required in validations we have to select a
value for it. Previously it was "just working" because the field
defaulted to the first option in the select.

We have to update the values we look for since in the interface we're
picking `"Energy"`, but the data we send out is `["energy"]`. The
approach I've used will only work for single word values -- that's good
enough for now.