Rename republish document service to bulk_publish
This service doesn't really perform a "republish" operation. It's used
for publishing draft, historic, imported documents.

Leaving it labelled as a republish service may encourage people to use
it for republishing the live version of a document, which could have
unexpected consequences.

Also remove optional argument from republish method. All callers of this
method pass `true` for the boolean, so get rid of the slightly confusing
optional argument altogether.