Move notify_listeners call out of publish method
The `notify_listeners` method is called by the `call` method in all
services except for the two publish services. As it was just a tail call
from the `publish` methods, moving it up makes it consistent with the
other services.

This had the added benefit of revealing the inconsistency between the
publish services and the other services -- the publish services notify
listeners *before* persisting the record; all other services notify
listeners *after* persisting the record. Addressing that inconsistency
will need to be picked up as part of a future commit, as it may affect
other parts of the system.