Add missing MAIB and RAIB cukes
The cucumber features for creating and editing MAIB and RAIB reports
never included the same level of scenarios as the other document
formats.

The step definitions for all of these missing scenarios already exist,
so we just need to add the scenarios to the feature file and tweak a
couple of steps.