Have the registry set the org slug attribute
The OrganisationalManualServiceRegistry takes an organisation slug and uses it
to build an organisational manual repository.

The registry caches that organisation slug.

The registry's sole user is the ManualsController. Before passing its
user-supplied params to the registry's services, ManualsController merges that
same organisation slug into its params.

This might not be the right place for this, but it is one approach to making it
nicer to use these 'organisational' services outside of the ManualsController
with less repetition of the configuration.