CliManualSectionRemover and features
CliManualSectionRemover uses Specialist Publisher's services, outputs feedback
to STDOUT/STDERR, accepts a user confirmation and generally encapsulates a lot
of what a bin script needs to do to remove a section from a manual.

This is hopefully a good interface for testing the script's behaviour
end-to-end as well as allowing us to keep the more procedural script file as
small as possible.

As it stands, to perform its task, the CliManualSectionRemover requires the ID
of the parent manual, the ID of the section and the slug of the owning
organisation. I'm hoping that last requirement can be refactored out shortly,
since I assume Specialist Publisher's services require an organisation slug
more for permission checking when being used from the controller than any real
need that would affect this administrative use case.

Some of the Cucumber steps added in this commit have terrible names. This is in
part due to having to avoid collisions with the existing features that require
manuals/etc to exist and go about creating them through the UI. With any luck
this can be fixed during some cleanup of Specialist Publisher's features at
a later date.