`require` reorder_manual_documents_service
The ReorderManualDocumentsService was the only service used in this registry
that wasn't already `require`d at the top of the file.

This was motivated by a desire to fix that inconsistency rather than noticing
any problems related to it.

There are other non-service classes used in this file that aren't `require`d,
but that's true in a lot of places throughout the codebase and I don't fully
understand the implications to know where to start resolving that.