Don't have the registry set the org slug attribute
This reverts commit c8c7898bbb21670110f8476e28edf587ab4eab10 and brings uses of
the OrganisationalManualServiceRegistry that're outside of the
ManualsController (i.e. in test helpers) in line with the controller's use.

Upcoming changes to access controls for GDS Editors (allowing them to access
Manuals outside of their own organisation) will involve the use of the
non-organisational services from within the controller, meaning right now it
makes more sense to apply the organisation_slug in the controller after all.

Besides that, we're no longer intending to use the organisational services
outside of the controller (as we were when this change first happened, when the
thinking was that this change would make organisational services more
convenient to use in those other contexts).

I also reckon this particular difference between the 2 types of
ManualServiceRegistry is at least for now making refactoring more difficult.